Move Banner logic out of HomeFragment
Banner related code could be in another class to reduce the responsibility of HomeFragment
This also helps if we want to have banner elsewhere.
